backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.40.2
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.8.2
PyYAML==6.0.2
markdown==3.6
unstructured==0.15.7
pdfminer.six==20231228
pymilvus==2.4.7
rank-bm25==0.2.2
jieba==0.42.1

# 深度文档理解
pdfplumber==0.11.4      # PDF 表格提取
python-docx==1.1.2      # Word 文档解析
openpyxl==3.1.5         # Excel 文件解析
pillow==10.4.0          # 图片处理
pytesseract==0.3.13     # OCR 文字识别（需要安装 Tesseract）
